requirements.txt
asgiref==3.6.0
Django==4.2.2
django-admin-rangefilter==0.10.0
django-debug-toolbar==4.1.0
django-extensions==3.2.3
django-filter==2.2.0
djangorestframework==3.15.2
importlib-metadata==6.6.0
Markdown==3.1.1
markdown-table==2019.4.13
public==2020.12.3
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
zipp==3.15.0
django-select2==8.1.2
django-admin-interface==0.26.1
django-admin-easy==0.7.0